Moving FeaturedCard to own file
