+++ src/views/homepage/components/featured-content-grid/featured-card.tsx
@@ -1,0 +1,25 @@
+import classNames from 'classnames'
+import { IconArrowRight24 } from '@hashicorp/flight-icons/svg-react/arrow-right-24'
+import CardLink from 'components/card-link'
+import Text from 'components/text'
+import s from './featured-card.module.css'
+
+const FeaturedCard = ({ children, className, href, title }: $TSFixMe) => {
+	return (
+		<CardLink
+			ariaLabel={title}
+			className={classNames(s.card, className)}
+			href={href}
+		>
+			<div>{children}</div>
+			<div className={s.learnMoreCta}>
+				<Text size={300} weight="medium">
+					Learn More
+				</Text>
+				<IconArrowRight24 />
+			</div>
+		</CardLink>
+	)
+}
+
+export default FeaturedCard
